requirements.txt
asgiref==3.8.1
Django==5.1.7
django-crispy-forms==2.3
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
